flaskLearn/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
numpy~=1.23.5
tensorflow~=2.12.0
Pillow~=9.5.0